Doing some code cleanup Fixing few UI issues slots are not geeting updated when local node does not have slots, Adding system activate feature
