+++ package.json
@@ -464,7 +464,8 @@
         "path-browserify": "1.0.1",
         "portfinder": "1.0.32",
         "sax-ts": "1.2.13",
-        "xml-js": "1.6.11"
+        "xml-js": "1.6.11",
+        "@vscode/codicons": "0.0.36"
     },
     "optionalDependencies": {
         "@tektronix/kic-cli-darwin-arm64": "0.19.3",
